Initial commit for refactoring visualization methods
- Extracted static method for creating *PapyrusVisualizer* from *TxtUMLExporter*, as it would violate function separation.

- Extracted cleanup operation from *TxtUMLExporter* to abstract method in *TxtUMLVisualizeWizard*, as it has to be different for different visualization methods anyway.

- Removed `outputFolder` property from *TxtUMLExporter*.

- **moved classes to packages**:
	- hu.elte.txtuml.export.papyrus.layout.txtuml.TxtUMLExporter
		- to: *hu.elte.txtuml.layout.export*
	- hu.elte.txtuml.export.papyrus.layout.txtuml.TxtUMLLayoutDescriptor
		- to: *hu.elte.txtuml.layout.export*
	- hu.elte.txtuml.export.papyrus.layout.txtuml.LayoutExportException
		- to: *hu.elte.txtuml.layout.export*
	- hu.elte.txtuml.export.papyrus.utils.LayoutUtils
		- to: *hu.elte.txtuml.layout.export.utils*
	- hu.elte.txtuml.export.papyrus.elementsarrangers.txtumllayout.LayoutVisualizerManager
		- to: *hu.elte.txtuml.visualizer*
	- hu.elte.txtuml.export.papyrus.elementsarrangers.ArrangeException
		- to: *hu.elte.txtuml.visualizer.exceptions*
