+++ dev/plugins/hu.elte.txtuml.export.papyrus/META-INF/MANIFEST.MF
@@ -49,5 +49,6 @@ Export-Package: hu.elte.txtuml.export.papyrus,
  hu.elte.txtuml.export.papyrus.papyrusmodelmanagers,
  hu.elte.txtuml.export.papyrus.preferences,
  hu.elte.txtuml.export.papyrus.wizardz
-Import-Package: hu.elte.txtuml.utils.diagrams,
+Import-Package: hu.elte.txtuml.export.papyrus,
+ hu.elte.txtuml.utils.diagrams,
  hu.elte.txtuml.utils.eclipse
