+++ dev/plugins/hu.elte.txtuml.layout.export/src/hu/elte/txtuml/layout/export/TxtUMLLayoutDescriptor.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.export.papyrus.layout.txtuml;
+package hu.elte.txtuml.layout.export;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -6,7 +6,6 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-import hu.elte.txtuml.layout.export.DiagramExportationReport;
 import hu.elte.txtuml.utils.Pair;
 
 /**
