+++ dev/plugins/hu.elte.txtuml.utils.eclipse/src/hu/elte/txtuml/utils/eclipse/wizards/TxtUMLVisualizeWizard.java
@@ -6,12 +6,17 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+
+import org.eclipse.core.runtime.CoreException;
+import java.io.IOException;
 
 import org.eclipse.jdt.core.IType;
 import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.jface.wizard.Wizard;
+import org.eclipse.swt.widgets.Display;
 
 import hu.elte.txtuml.utils.Logger;
 import hu.elte.txtuml.utils.Pair;
@@ -85,6 +90,15 @@ public abstract class TxtUMLVisualizeWizard extends Wizard {
 				layoutConfigs.values().stream().flatMap(c -> c.stream())
 				.map(layout -> layout.getJavaProject().getElementName()).collect(Collectors.toList()));
 
+		Display.getDefault().syncExec(() -> {
+			try {
+				cleanBeforeVisualization(layoutConfigs.keySet());
+			} catch (CoreException | IOException e) {
+				Dialogs.errorMsgb("txtUML export Error - cleaning resources",
+						"Error occured when cleaning resources.", e);
+			}
+		});
+		
 		return exportDiagrams(layoutConfigs, txtUMLLayout);
 	}
 
@@ -94,6 +108,28 @@ public abstract class TxtUMLVisualizeWizard extends Wizard {
 	 * Abstract method for diagram export.
 	 */
 	protected abstract boolean exportDiagrams(Map<Pair<String, String>, List<IType>> layoutConfigs, List<IType> txtUMLLayout);
+	
+	/**
+	 * Abstract method for cleanup.
+	 * 
+	 * Performs cleanup according to the visualization method used.
+	 * 
+	 * @param layouts
+	 * 			the modelName, projectName pairs to be cleaned before export
+	 * 
+	 * @throws CoreException
+	 *             if this method fails. Reasons include:
+	 *             <ul>
+	 *             <li>This resource could not be deleted for some reason.</li>
+	 *             <li>This resource or one of its descendants is out of sync
+	 *             with the local file system and force is false.</li>
+	 *             <li>Resource changes are disallowed during certain types of
+	 *             resource change event notification. See IResourceChangeEvent
+	 *             for more details.</li>
+	 *             </ul>
+	 * 
+	 */
+	protected abstract void cleanBeforeVisualization(Set<Pair<String, String>> layouts) throws CoreException, IOException ;
 
 	protected void checkNoLayoutDescriptionsSelected() throws InterruptedException {
 		if (selectTxtUmlPage.getTxtUmlLayouts().isEmpty()) {
